requirements.txt
asgiref==3.7.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-paypal==2.0
idna==3.4
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.6.1
urllib3==2.0.2
